+++ packages/nimble-components/src/icon-base/index.ts
@@ -1,13 +1,23 @@
+import { attr } from '@microsoft/fast-element';
 import { DesignSystem, FoundationElement } from '@microsoft/fast-foundation';
 import type { NimbleIcon } from '@ni/nimble-tokens/dist/icons/js';
 import { template } from './template';
 import { styles } from './styles';
+import type { IconSeverity } from './types';
 
 /**
  * The base class for icon components
  */
 export class Icon extends FoundationElement {
-    public constructor(public icon: NimbleIcon) {
+    /**
+     * @public
+     * @remarks
+     * HTML Attribute: severity
+     */
+    @attr
+    public severity: IconSeverity;
+
+    public constructor(/** @internal */ public readonly icon: NimbleIcon) {
         super();
     }
 }