+++ packages/nimble-components/src/tooltip/tests/tooltip.spec.ts
@@ -144,7 +144,7 @@ describe('Tooltip', () => {
 
     it('should render the default state when selected and not render an icon when true', async () => {
         element.visible = true;
-        element.classList.add('icon-visible');
+        element.iconVisible = true;
 
         await connect();
         await DOM.nextUpdate();
@@ -155,9 +155,9 @@ describe('Tooltip', () => {
         await disconnect();
     });
 
-    it('should render the error state when selected', async () => {
+    it('should render the error severity when selected', async () => {
         element.visible = true;
-        element.classList.add('fail');
+        element.severity = 'error';
 
         await connect();
         await DOM.nextUpdate();
@@ -168,10 +168,10 @@ describe('Tooltip', () => {
         await disconnect();
     });
 
-    it('should render the error state when selected and render the corresponding icon when true', async () => {
+    it('should render the error severity when selected and render the corresponding icon when true', async () => {
         element.visible = true;
-        element.classList.add('fail');
-        element.classList.add('icon-visible');
+        element.severity = 'error';
+        element.iconVisible = true;
 
         await connect();
         await DOM.nextUpdate();
@@ -182,9 +182,9 @@ describe('Tooltip', () => {
         await disconnect();
     });
 
-    it('should render the info state when selected', async () => {
+    it('should render the information severity when selected', async () => {
         element.visible = true;
-        element.classList.add('information');
+        element.severity = 'information';
 
         await connect();
         await DOM.nextUpdate();
@@ -195,10 +195,10 @@ describe('Tooltip', () => {
         await disconnect();
     });
 
-    it('should render the information state when selected and render the corresponding icon when true', async () => {
+    it('should render the information severity when selected and render the corresponding icon when true', async () => {
         element.visible = true;
-        element.classList.add('information');
-        element.classList.add('icon-visible');
+        element.severity = 'information';
+        element.iconVisible = true;
 
         await connect();
         await DOM.nextUpdate();