+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Thu, 15 Sep 2022 18:33:47 GMT and should not be manually modified.
+This log was last generated on Thu, 29 Sep 2022 20:03:25 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 9.0.0
+
+Thu, 29 Sep 2022 20:03:25 GMT
+
+### Major changes
+
+- Migrated from CSS classes to element attributes for configuring components. For the list of specific elements and classes changed see: [#738](https://github.com/ni/nimble/pull/738). ([ni/nimble@9023d71](https://github.com/ni/nimble/commit/9023d71ddb4740a624e5b9eac0114a0c91fbca78))
+- Bump @ni/nimble-components to v12.0.0
+
+## 8.5.0
+
+Tue, 27 Sep 2022 21:39:15 GMT
+
+### Minor changes
+
+- Angular integration for radio button and group ([ni/nimble@6906246](https://github.com/ni/nimble/commit/6906246ea159f42edcd34b2cc881611afa32186d))
 
 ## 8.4.12